@@ -9,14 +9,18 @@ import (
 	"github.com/spf13/cobra"
 )
 
-// version is the current version of this program.
-const version = "1.6.0"
+const version = "1.7.1"
+
+var (
+	BuildTime = "unknown"
+	GitSha1   = "dev"
+)
 
 var versionCmd = &cobra.Command{
 	Use:   "version",
 	Short: "Show version of this program and of Lilypond",
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Printf("%s v%s %s/%s\n", cmd.Root().Name(), version, runtime.GOOS, runtime.GOARCH)
+		fmt.Printf("%s v%s-%s built %s %s/%s\n", cmd.Root().Name(), version, GitSha1, BuildTime, runtime.GOOS, runtime.GOARCH)
 		fmt.Println(lilyVersion())
 	},
 }
